+++ pojo/src/main/java/com/fengwei/pojo/ProductSsd.java
@@ -3,6 +3,7 @@ package com.fengwei.pojo;
 import javax.persistence.Column;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Table(name = "product_ssd")
@@ -17,7 +18,7 @@ public class ProductSsd {
     private String model;
 
     @Column(name = "form_factor")
-    private Long formFactor;
+    private BigDecimal formFactor;
 
     private Integer capacity;
 
@@ -34,16 +35,16 @@ public class ProductSsd {
     @Column(name = "max_write")
     private Integer maxWrite;
 
-    private Long height;
+    private BigDecimal height;
 
-    private Long width;
+    private BigDecimal width;
 
-    private Long depth;
+    private BigDecimal depth;
 
     @Column(name = "main_image")
     private String mainImage;
 
-    private Long price;
+    private BigDecimal price;
 
     private Integer stock;
 
@@ -119,14 +120,14 @@ public class ProductSsd {
     /**
      * @return form_factor
      */
-    public Long getFormFactor() {
+    public BigDecimal getFormFactor() {
         return formFactor;
     }
 
     /**
      * @param formFactor
      */
-    public void setFormFactor(Long formFactor) {
+    public void setFormFactor(BigDecimal formFactor) {
         this.formFactor = formFactor;
     }
 
@@ -217,42 +218,42 @@ public class ProductSsd {
     /**
      * @return height
      */
-    public Long getHeight() {
+    public BigDecimal getHeight() {
         return height;
     }
 
     /**
      * @param height
      */
-    public void setHeight(Long height) {
+    public void setHeight(BigDecimal height) {
         this.height = height;
     }
 
     /**
      * @return width
      */
-    public Long getWidth() {
+    public BigDecimal getWidth() {
         return width;
     }
 
     /**
      * @param width
      */
-    public void setWidth(Long width) {
+    public void setWidth(BigDecimal width) {
         this.width = width;
     }
 
     /**
      * @return depth
      */
-    public Long getDepth() {
+    public BigDecimal getDepth() {
         return depth;
     }
 
     /**
      * @param depth
      */
-    public void setDepth(Long depth) {
+    public void setDepth(BigDecimal depth) {
         this.depth = depth;
     }
 
@@ -273,14 +274,14 @@ public class ProductSsd {
     /**
      * @return price
      */
-    public Long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
     /**
      * @param price
      */
-    public void setPrice(Long price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
